+++ packages/eslint-plugin-stories/src/rules/__tests__/noComponentsWithoutStoryTestComponent.stories.tsx
